apps/agente/requirements.txt
google-adk>=0.6.0
pytz